Refactor Tools page into modular components with custom hooks
